requirements.txt
Django==2.2.1
django-constance==2.4.0
django-picklefield==2.0
django-rest-framework==0.1.0
django-safedelete==0.5.1
django-smart-selects==1.5.4
django-suit==2.0a1
django-super-inlines==0.1.4
django-tinymce==2.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
mysqlclient==1.4.4
Pillow==6.0.0
python-monkey-business==1.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
